Rename org.bdgenomics.adam.rdd.variation package to o.b.a.rdd.variant
